Made blueprint a default plugin that can be disabled by bool on init
