+++ flask_nemo/data/templates/container.html
@@ -28,19 +28,7 @@
       {% include "main::menu.html" %}
     </nav>
     <div class="col-md-10" id="reading-container">
-        <nav id="breadcrumb">
-            <ol class="breadcrumb">
-              <li><a href="{{url_for(".r_index")}}">Home</a></li>
-              {% for crumb in breadcrumbs: %}
-                {%if crumb["link"] != None %}
-           
-                  <li><a href="{{url_for(crumb["link"],**crumb["args"])}}">{{crumb["title"]}}</a></li>
-                {% else %}
-                  <li class="active">{{crumb["title"]}}</li>
-                {% endif %}
-              {% endfor %}
-            </ol>
-        </nav>
+          {% include "main::breadcrumb.html" %}
           {% block article %}
           {% endblock %}
     </div>
