+++ flask_nemo/data/templates_plugins/breadcrumb/breadcrumb.html
@@ -1,0 +1,13 @@
+<nav id="breadcrumb">
+    <ol class="breadcrumb">
+      <li><a href="{{url_for(".r_index")}}">Home</a></li>
+      {% for crumb in breadcrumbs: %}
+        {%if crumb["link"] != None %}
+
+          <li><a href="{{url_for(crumb["link"],**crumb["args"])}}">{{crumb["title"]}}</a></li>
+        {% else %}
+          <li class="active">{{crumb["title"]}}</li>
+        {% endif %}
+      {% endfor %}
+    </ol>
+</nav>
