+++ tests/test_routes.py
@@ -5,6 +5,7 @@
 from .resources import NemoResource
 from .test_controller import NemoTestControllers
 from flask_nemo import Nemo
+from flask_nemo.default import Breadcrumb
 from flask import Markup, Flask
 from lxml import etree
 from mock import Mock, patch, call
@@ -226,9 +227,9 @@ class NemoTestRoutes(NemoResource):
         """
         with patch("requests.get", return_value=self.getCapabilities):
             with patch("flask_nemo.render_template") as patched:
-                self.nemo.render("index.html", test="123", value="value", url={})
+                self.nemo.render("main::index.html", test="123", value="value", url={})
                 patched.assert_called_once_with(
-                    "index.html",
+                    "main::index.html",
                     collections={'latinLit', 'greekLit'},
                     test="123",
                     value="value",
@@ -243,7 +244,7 @@ class NemoTestRoutes(NemoResource):
         """
         with patch("requests.get", return_value=self.getCapabilities):
             with patch("flask_nemo.render_template") as patched:
-                self.nemo.render(
+                data = self.nemo.render(
                     "index.html",
                     test="123",
                     value="value",
@@ -351,72 +352,3 @@ class NemoTestRoutes(NemoResource):
 
         with self.assertRaises(TemplateNotFound):
             html, path, function = blueprint.jinja_loader.get_source("", "addendum::unknown.html")
-
-    def test_make_passage_breadcrumb(self):
-        """ passage breadcrumb should include all components up to passage and passage not linked
-        """
-        with patch("requests.get", return_value=self.getCapabilities):
-            bc = self.nemo.make_breadcrumbs(
-                textgroups=self.nemo.get_textgroups(),
-                version=self.nemo.get_text("latinLit", "phi1294", "phi002", "perseus-lat2"),
-                lang="eng",
-                url={
-                    "collection": "latinLit",
-                    "textgroup": "phi1294",
-                    "work": "phi002",
-                    "version": "perseus-lat2",
-                    "passage_identifier": "1.1"
-                    })
-            self.assertEqual(bc,[
-                {'link': '.r_collection', 'title': 'latinLit', 'args': {'collection': 'latinLit'}},
-                {'link': '.r_texts', 'title': 'Martial', 'args': {'textgroup': 'phi1294', 'collection': 'latinLit'}},
-                {'link': '.r_version', 'title': 'Epigrammata Label', 'args': {'textgroup': 'phi1294', 'collection': 'latinLit', 'work':'phi002','version':'perseus-lat2'}},
-                {'link': None, 'title': '1.1', 'args': {'textgroup': 'phi1294', 'collection': 'latinLit', 'work':'phi002','version':'perseus-lat2', 'passage_identifier':"1.1"}}
-            ])
-
-    def test_make_version_breadcrumb(self):
-        """ version breadcrumb should include all components up to version and version not linked
-        """
-        with patch("requests.get", return_value=self.getCapabilities):
-            bc = self.nemo.make_breadcrumbs(
-                textgroups=self.nemo.get_textgroups(),
-                version = self.nemo.get_text("latinLit","phi1294","phi002","perseus-lat2"),
-                lang="eng",
-                url={
-                    "collection": "latinLit",
-                    "textgroup": "phi1294",
-                    "work": "phi002",
-                    "version": "perseus-lat2"
-                    })
-            self.assertEqual(bc,[
-                {'link': '.r_collection', 'title': 'latinLit', 'args': {'collection': 'latinLit'}},
-                {'link': '.r_texts', 'title': 'Martial', 'args': {'textgroup': 'phi1294', 'collection': 'latinLit'}},
-                {'link': None, 'title': 'Epigrammata Label', 'args': {'textgroup': 'phi1294', 'collection': 'latinLit', 'work':'phi002','version':'perseus-lat2'}}
-            ])
-
-    def test_make_textgroup_breadcrumb(self):
-        """ textgroup breadcrumb should include all components up to textgroup and textgroup not linked
-        """
-        with patch("requests.get", return_value=self.getCapabilities):
-            bc = self.nemo.make_breadcrumbs(
-                textgroups=self.nemo.get_textgroups(),
-                lang="eng",
-                url={
-                    "collection": "latinLit",
-                    "textgroup": "phi1294",
-                    })
-            self.assertEqual(bc,[
-                {'link': '.r_collection', 'title': 'latinLit', 'args': {'collection': 'latinLit'}},
-                {'link': None, 'title': 'Martial', 'args': {'textgroup': 'phi1294', 'collection': 'latinLit'}}
-            ])
-
-    def test_make_collection_breadcrumb(self):
-        """ collection breadcrumb should include only collection not linked
-        """
-        with patch("requests.get", return_value=self.getCapabilities):
-            bc = self.nemo.make_breadcrumbs(
-                lang="eng",
-                url={ "collection": "latinLit"})
-            self.assertEqual(bc,[
-                {'link': None, 'title': 'latinLit', 'args': {'collection': 'latinLit'}}
-            ])
